requirements.txt
Django==4.1
torch==1.9.0
Pillow==8.3.2
